+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+* Added a new `VpnConnectAction`: `VPN_CA_REJECT`. 
+
 ## 0.92.107
 
 * Added `VpnConnectedInfo::relay_address`.